+++ amttest/errors/badrequest.py
@@ -1,26 +1,19 @@
 from .apierror import APIError
 
 from flask import jsonify, make_response, Blueprint
-from werkzeug.exceptions import BadRequest
 
 BAD_REQUEST_BP = Blueprint('BadRequest', __name__)
 
-class AmtBadRequest(Exception):
+class BadRequest(APIError):
 
     def __init__(self, message, **kwargs):
-        #super().__init__(message, **kwargs)
-        self.message = message
-        if kwargs:
-            self.__dict__.update(kwargs)
+        super().__init__(message, **kwargs)
         self.code = 400
         self.error = 'Bad Request'
 
-    def to_json(self):
-        return jsonify(self.__dict__)
 
-
-@BAD_REQUEST_BP.app_errorhandler(AmtBadRequest)
+@BAD_REQUEST_BP.app_errorhandler(BadRequest)
 def handle_bad_request(message, **kwargs):
-    error = AmtBadRequest(message, **kwargs)
+    error = BadRequest(message, **kwargs)
     print(error.to_json())
     return make_response(error.to_json())
